smekaylo/requirements.txt
anyio==4.4.0
certifi==2024.7.4
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-telegram-bot==21.4
sniffio==1.3.1
TgCrypto==1.2.5
anyio==4.4.0
certifi==2024.7.4
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
psycopg2-binary==2.9.9
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-telegram-bot==21.4
sniffio==1.3.1
TgCrypto==1.2.5
typing_extensions==4.12.2
